// src/app/enums/nutrition.enum.ts
export enum Nutrition {
  maxDailyEnergyKJ = "Energia KJ",
  maxDailyEnergyKcal = "Energia Kcal",
  maxDailyProtein = "Fehérje",
  maxDailyFat = "Zsír",
  maxDailySaturatedFat = "Telített zsír",
  maxDailyCarbohydrate = "Szénhidrát",
  maxDailySugar = "Cukor",
  maxGlycemicIndexPerMeal = "Glikémiás index/étkezés",
  maxDailyFibre = "Rost",
  maxDailyNatrium = "Nátrium",
  maxDailyPotassium = "Kálium",
  maxDailyCalcium = "Kalcium",
  maxDailyMagnesium = "Magnézium"
}